make it somewhat work
