+++ gradia/graphics/gradient_gen.c
@@ -1,39 +1,55 @@
-// Copyright (C) 2025 Alexander Vanhee
-//
-// This program is free software: you can redistribute it and/or modify it
-// under the terms of the GNU General Public License as published by the Free
-// Software Foundation, either version 3 of the License, or (at your option)
-// any later version.
-//
-// This program is distributed in the hope that it will be useful, but WITHOUT
-// ANY WARRANTY; without even the implied warranty of  MERCHANTABILITY or
-// FITNESS FOR A PARTICULAR PURPOSE. See the GNU General Public License for
-// more details.
-//
-// You should have received a copy of the GNU General Public License along with
-// this program.  If not, see <http://www.gnu.org/licenses/>.
-
 #include <math.h>
 #include <stdint.h>
 
-void generate_gradient(uint8_t* pixels, int width, int height,
-                      int start_r, int start_g, int start_b,
-                      int end_r, int end_g, int end_b,
-                      double angle) {
+typedef struct {
+    double position; // 0.0 to 1.0
+    uint8_t r, g, b;
+} ColorStop;
+
+static void interpolate_color(double t, const ColorStop* stops, int num_stops, uint8_t* out_r, uint8_t* out_g, uint8_t* out_b) {
+    if (t <= stops[0].position) {
+        *out_r = stops[0].r;
+        *out_g = stops[0].g;
+        *out_b = stops[0].b;
+        return;
+    }
+    for (int i = 1; i < num_stops; i++) {
+        if (t <= stops[i].position) {
+            double ratio = (t - stops[i-1].position) / (stops[i].position - stops[i-1].position);
+            *out_r = (uint8_t)(stops[i-1].r + (stops[i].r - stops[i-1].r) * ratio);
+            *out_g = (uint8_t)(stops[i-1].g + (stops[i].g - stops[i-1].g) * ratio);
+            *out_b = (uint8_t)(stops[i-1].b + (stops[i].b - stops[i-1].b) * ratio);
+            return;
+        }
+    }
+
+    *out_r = stops[num_stops - 1].r;
+    *out_g = stops[num_stops - 1].g;
+    *out_b = stops[num_stops - 1].b;
+}
+
+void generate_gradient(
+    uint8_t* pixels, int width, int height,
+    const ColorStop* stops, int num_stops,
+    double angle, int mode // 0 = linear, 1 = conic
+) {
     double cos_angle = cos(angle * M_PI / 180.0);
     double sin_angle = sin(angle * M_PI / 180.0);
 
-    double corners[4];
-    corners[0] = 0 * cos_angle + 0 * sin_angle;
-    corners[1] = (width-1) * cos_angle + 0 * sin_angle;
-    corners[2] = 0 * cos_angle + (height-1) * sin_angle;
-    corners[3] = (width-1) * cos_angle + (height-1) * sin_angle;
+    double min_coord = 0.0, max_coord = 1.0;
+    if (mode == 0) {
+        double corners[4];
+        corners[0] = 0 * cos_angle + 0 * sin_angle;
+        corners[1] = (width-1) * cos_angle + 0 * sin_angle;
+        corners[2] = 0 * cos_angle + (height-1) * sin_angle;
+        corners[3] = (width-1) * cos_angle + (height-1) * sin_angle;
 
-    double min_coord = corners[0];
-    double max_coord = corners[0];
-    for (int i = 1; i < 4; i++) {
-        if (corners[i] < min_coord) min_coord = corners[i];
-        if (corners[i] > max_coord) max_coord = corners[i];
+        min_coord = corners[0];
+        max_coord = corners[0];
+        for (int i = 1; i < 4; i++) {
+            if (corners[i] < min_coord) min_coord = corners[i];
+            if (corners[i] > max_coord) max_coord = corners[i];
+        }
     }
 
     double range = max_coord - min_coord;
@@ -41,18 +57,27 @@ void generate_gradient(uint8_t* pixels, int width, int height,
 
     for (int y = 0; y < height; y++) {
         for (int x = 0; x < width; x++) {
-            double coord = x * cos_angle + y * sin_angle;
-            double t = (coord - min_coord) / range;
+            double t;
+            if (mode == 0) { // Linear
+                double coord = x * cos_angle + y * sin_angle;
+                t = (coord - min_coord) / range;
+            } else { // Conic
+                double dx = x - width / 2.0;
+                double dy = y - height / 2.0;
+                t = (atan2(dy, dx) + M_PI) / (2 * M_PI);
+            }
 
-            // Clamp t to [0, 1]
             if (t < 0) t = 0;
             if (t > 1) t = 1;
 
+            uint8_t r, g, b;
+            interpolate_color(t, stops, num_stops, &r, &g, &b);
+
             int idx = (y * width + x) * 4;
-            pixels[idx] = (uint8_t)(start_r + (end_r - start_r) * t);     // R
-            pixels[idx + 1] = (uint8_t)(start_g + (end_g - start_g) * t); // G
-            pixels[idx + 2] = (uint8_t)(start_b + (end_b - start_b) * t); // B
-            pixels[idx + 3] = 255;                                        // A
+            pixels[idx] = r;
+            pixels[idx + 1] = g;
+            pixels[idx + 2] = b;
+            pixels[idx + 3] = 255;
         }
     }
 }
