+++ gradia/ui/widget/gradient_preset_button.py
@@ -1,8 +1,24 @@
+# Copyright (C) 2025 Alexander Vanhee, tfuxu
+#
+# This program is free software: you can redistribute it and/or modify
+# it under the terms of the GNU General Public License as published by
+# the Free Software Foundation, either version 3 of the License, or
+# (at your option) any later version.
+#
+# This program is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+# GNU General Public License for more details.
+#
+# You should have received a copy of the GNU General Public License
+# along with this program.  If not, see <https://www.gnu.org/licenses/>.
+#
+# SPDX-License-Identifier: GPL-3.0-or-later
 from gi.repository import Adw, Gtk
 from typing import Callable, Optional
 from gi.repository import Gtk, GObject
 from gradia.utils.colors import HexColor
-from gradia.graphics.gradient import GradientBackground
+from gradia.graphics.gradient import GradientBackground, Gradient
 from gradia.app_constants import PREDEFINED_GRADIENTS
 
 class GradientPresetButton(Gtk.MenuButton):
@@ -10,17 +26,14 @@ class GradientPresetButton(Gtk.MenuButton):
 
     def __init__(
         self,
-        callback: Optional[Callable[[HexColor, HexColor, int], None]] = None,
+        callback: Optional[Callable[[Gradient], None]] = None,
         **kwargs
     ) -> None:
         super().__init__(**kwargs)
-
         self.callback = callback
-
         self.set_icon_name("view-grid-symbolic")
         self.set_valign(Gtk.Align.CENTER)
         self.add_css_class("flat")
-
         self._setup_popover()
 
     def _setup_popover(self) -> None:
@@ -47,9 +60,8 @@ class GradientPresetButton(Gtk.MenuButton):
         self.popover.set_child(main_box)
 
     def _create_gradient_buttons(self) -> None:
-        for i, (start, end, angle) in enumerate(PREDEFINED_GRADIENTS[:6]):
+        for i, gradient in enumerate(PREDEFINED_GRADIENTS[:6]):
             gradient_name = f"gradient-preset-{i}"
-
             button = Gtk.Button(
                 name=gradient_name,
                 focusable=True,
@@ -58,20 +70,16 @@ class GradientPresetButton(Gtk.MenuButton):
                 height_request=40
             )
             button.add_css_class("gradient-preset")
-
-            self._apply_gradient_to_button(button, start, end, angle)
-
-            button.connect("clicked", self._on_gradient_button_clicked, start, end, angle)
-
+            self._apply_gradient_to_button(button, gradient)
+            button.connect("clicked", self._on_gradient_button_clicked, gradient)
             self.flowbox.append(button)
 
-    def _apply_gradient_to_button(self, button: Gtk.Button, start: str, end: str, angle: int) -> None:
+    def _apply_gradient_to_button(self, button: Gtk.Button, gradient: Gradient) -> None:
         css = f"""
             button.gradient-preset {{
-                background-image: linear-gradient({angle}deg, {start}, {end});
+                background-image: {gradient.to_css()};
             }}
         """
-
         css_provider = Gtk.CssProvider()
         css_provider.load_from_string(css)
         button.get_style_context().add_provider(
@@ -79,25 +87,25 @@ class GradientPresetButton(Gtk.MenuButton):
             Gtk.STYLE_PROVIDER_PRIORITY_APPLICATION + 3
         )
 
-    def _on_gradient_button_clicked(self, button: Gtk.Button, start: HexColor, end: HexColor, angle: int) -> None:
+    def _on_gradient_button_clicked(self, button: Gtk.Button, gradient: Gradient) -> None:
         self.popover.popdown()
-
         if self.callback:
-            self.callback(start, end, angle)
+            self.callback(gradient)
 
-    def set_callback(self, callback: Callable[[HexColor, HexColor, int], None]) -> None:
+    def set_callback(self, callback: Callable[[Gradient], None]) -> None:
         self.callback = callback
 
-    def set_gradient_presets(self, gradients: list[tuple[HexColor, HexColor, int]]) -> None:
+    def set_gradient_presets(self, gradients: list[Gradient]) -> None:
+        # Clear existing buttons
         child = self.flowbox.get_first_child()
         while child:
             next_child = child.get_next_sibling()
             self.flowbox.remove(child)
             child = next_child
 
-        for i, (start, end, angle) in enumerate(gradients[:6]):
+        # Add new gradient buttons
+        for i, gradient in enumerate(gradients[:6]):
             gradient_name = f"gradient-preset-{i}"
-
             button = Gtk.Button(
                 name=gradient_name,
                 focusable=True,
@@ -106,8 +114,6 @@ class GradientPresetButton(Gtk.MenuButton):
                 height_request=40
             )
             button.add_css_class("gradient-preset")
-
-            self._apply_gradient_to_button(button, start, end, angle)
-            button.connect("clicked", self._on_gradient_button_clicked, start, end, angle)
-
+            self._apply_gradient_to_button(button, gradient)
+            button.connect("clicked", self._on_gradient_button_clicked, gradient)
             self.flowbox.append(button)
